Clone E869120/kyopro-tessoku
Clone E869120/kyopro-tessoku

Hi, I think that "kyopro-tessoku" is the very concise and elegant algorithm book for the beginners and advanced programmers to learn how efficient codes to write in. But it's found several faults in the codes of the repos. So, I want to verify the codes to know what is the proper answers in author's algorithm solutions and its real and true meanings.